+++ equinox-shipping/Watchdog/Infrastructure.fs
@@ -4,6 +4,11 @@ module Shipping.Watchdog.Infrastructure
 open Serilog
 open System
 
+module Log =
+
+    /// Allow logging to filter out emission of log messages whose information is also surfaced as metrics
+    let isStoreMetrics e = Filters.Matching.WithProperty("isMetric").Invoke e
+
 module EnvVar =
 
     let tryGet varName : string option = Environment.GetEnvironmentVariable varName |> Option.ofObj